Relocated config.py to tests directory, since both integration and unit tests used it
