+++ tests/integration_tests/test_user.py
@@ -3,7 +3,7 @@ import pytest
 
 from src.app import create_app
 
-from tests.integration_tests.config import (
+from tests.config import (
     configure_ports_without_auth,
     configure_ports_with_auth,
 )
