+++ tests/unit_tests/adapters/routers/test_inference_router.py
@@ -12,7 +12,7 @@ from core.ports.database_port import DatabasePort
 
 from src.app import create_app
 
-from tests.integration_tests.config import (
+from tests.config import (
     configure_ports_without_auth,
     configure_ports_with_auth,
 )
