+++ tests/unit_tests/adapters/routers/test_result_router.py
@@ -13,7 +13,7 @@ from core.model.result import Result, ResultCreation
 
 from tests.mocks.mongo_mock import MongoMock
 
-from tests.integration_tests.config import (
+from tests.config import (
     configure_ports_without_auth,
     configure_ports_with_auth,
 )
