+++ tests/unit_tests/adapters/routers/test_user_router.py
@@ -9,7 +9,7 @@ from core.model.user import User, UserCreationForm
 
 from src.app import create_app
 
-from tests.integration_tests.config import (
+from tests.config import (
     configure_ports_without_auth,
     configure_ports_with_auth,
 )
